NStrings

Documentation for NStrings.

NStrings.NStrings

NStrings.AbstractNMString

NStrings.AbstractNString

NStrings.CNMString

NStrings.CNString

NStrings.NMString

NStrings.NString

NStrings.NStrings — Module

NStrings.AbstractNMString — Type

AbstractNMString{N,M}

AbstractNString that represents a string of N codeunits but contains M codeunits.

NStrings.AbstractNString — Type

AbstractNString{N}

Represents a string of N codeunits

NStrings.CNMString — Type

CNMString{N,M}

AbstractNMString that represents a string of the first N codeunits but contained within a NTuple{M, UInt8} while containing no internal NULs having a terminal NUL at codeunit N+1.

NStrings.CNString — Type

CNString(data::NTuple{N,UInt8})

AbstractNString that wraps a NTuple{N,UInt8}.

NStrings.NMString — Type

NMString{N,M}

AbstractNMString that represents a string of the first N codeunits but contained within a NTuple{M, UInt8}

NStrings.NString — Type

NString(data::NTuple{N,UInt8})

AbstractNString that wraps a NTuple{N,UInt8}.