StaticStrings

Documentation for StaticStrings.

Fixed-length strings wrapping a NTuple for Julia.

Introduction

StaticStrings.jl implements several AbstractString subtypes that wrap a NTuple{N,UInt8}. An AbstractStaticString is a AbstractString with N codeunits.

The concrete subtypes of AbstractStaticString are as follows.

StaticString{N} is a wrapper of a NTuple{N,UInt8} of exactly N codeunits, padded with \0, nul, bytes.

SubStaticString{N, R <: AbstractUnitRange} is a wrapper of a NTuple{N,UInt8} of up to N codeunits, with a unit range indicating a subset of codeunits.

CStaticString{N} is similar to a StaticString but requires all the NUL bytes to be terminal codeunits. The struct also contains an extra terminal NUL.

PaddedStaticString{N,PAD} is siimlar to StaticString but is padded with an arbitrary byte codeunit.

Usage

julia> using StaticStrings

julia> Static"Hello world!"
Static"Hello world!"12

julia> Static"Hello world!" |> typeof
StaticString{12}

julia> CStatic"Hello world!"
CStatic"Hello world!"12

julia> ccall(:printf, Cint, (Ptr{Cchar},), cs)
Hello world!
13

julia> ccall(:printf, Cint, (Ptr{CStaticString{13}},), Ref(cs))
Hello world!
13

julia> Padded"Hello "20
Padded"Hello "20

julia> ps = Padded"Hello "20
Padded"Hello "20

julia> StaticString(ps)
Static"Hello               "20

julia> push!(StaticString["Hello"], "Bye")
2-element Vector{StaticString}:
 Static"Hello"5
 Static"Bye"5

StaticStrings.StaticStrings

StaticStrings.AbstractStaticString

StaticStrings.CStaticString

StaticStrings.PaddedStaticString

StaticStrings.StaticString

StaticStrings.SubStaticString

StaticStrings.data

StaticStrings.pad

StaticStrings.@CStatic_str

StaticStrings.@Padded_str

StaticStrings.@Static_str

StaticStrings.StaticStrings — Module

StaticStrings.jl

The StaticStrings.jl package implements an AbstractString based on NTuple{N,UInt8}.

julia> using StaticStrings

julia> Static"Thank you for using StaticStrings.jl"
Static"Thank you for using StaticStrings.jl"36

julia> Static"Thank you for using StaticStrings.jl"64
Static"Thank you for using StaticStrings.jl\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0"64

julia> StaticString((0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x20, 0x77, 0x6f, 0x72, 0x6c, 0x64))
Static"Hello world"11

See StaticString, SubStaticString, CStaticString, PaddedStaticString

StaticStrings.AbstractStaticString — Type

AbstractStaticString{N}

Represents a string of N codeunits

StaticStrings.StaticString — Type

StaticString(data::NTuple{N,UInt8})
Static"string"N

AbstractStaticString that stores codeunits in a NTuple{N,UInt8}.

StaticStrings.@Static_str — Macro

Static"string"[N]

Create a StaticString from "string". Optionally, specify the number of codeunits, N.

Examples

julia> Static"großartig"
Static"großartig"10

julia> Static"verehrungswürdig"20
Static"verehrungswürdig\0\0\0"20

StaticStrings.SubStaticString — Type

SubStaticString(data::NTuple{N, UInt8}, ind::Integer)
SubStaticString(data::NTuple{N, UInt8}, ind::AbstractUnitRange)
SubStatic"string"N

AbstractStaticString that stores up to N codeunits in a NTuple{N,UInt8}. The actual codeunits used are a subset indicated by an AbstractUnitRange.

Missing docstring.

Missing docstring for @SubStatic. Check Documenter's build log for details.

StaticStrings.CStaticString — Type

CStaticString(data::NTuple{N,UInt8})
CStatic"string"N

AbstractStaticString that stores codeunits in a NTuple{N,UInt8} but requires NUL codeunits to be at the end.

N.B. The size of a CStaticString{N} is N+1 bytes.

StaticStrings.@CStatic_str — Macro

CStatic"string"[N]

Create a CStaticString from "string". Optionally, specify the number of codeunits, N.

Examples

julia> CStatic"Orada mısın"
CStatic"Orada mısın"13

julia> CStatic"Orada mısın"25
CStatic"Orada mısın"25

julia> ccall(:jl_printf, Cint, (Ptr{Nothing}, Ptr{Cchar},), stdout isa IOContext ? stdout.io : stdout, CStatic"Orada mısın\n"25)
Orada mısın
14

StaticStrings.PaddedStaticString — Type

PaddedStaticString{N,PAD}(data::NTuple{N,UInt8})
Padded"string[PAD]"N

AbstractStaticString that is padded with PAD.

StaticStrings.@Padded_str — Macro

Padded"string[PAD]"N

Create a PaddedStaticString of N codeunits from "string". The last codeunit in the provided string becomes the PAD.

Examples

julia> Padded"私は元気です。 ありがとうございました。 "64
Padded"私は元気です。 ありがとうございました。 "64

julia> Padded"私は元気です。 ありがとうございました。 "64 |> StaticString
Static"私は元気です。 ありがとうございました。      "64

StaticStrings.data — Function

StaticStrings.data(string::AbstractStaticString{N})::NTuple{N,UInt8} where N

Retrieve the internal Tuple containing the N stored UInt8 code units.

StaticStrings.pad — Function

StaticStrings.pad(string::PaddedStaticString)

Retrieve the UInt8 code unit used for padding.

StaticStrings.pad(s::AbstractString, N::Integer, PAD::UInt8=0x0)::PaddedStaticString

Pad an AbstractString to N code units with the code unit PAD.