StaticStrings

Documentation for StaticStrings.

Fixed-length strings wrapping a NTuple for Julia.

Introduction

StaticStrings.jl implements several AbstractString subtypes that wrap a NTuple{N,UInt8}. An AbstractStaticString is a AbstractString with N codeunits.

The concrete subtypes of AbstractStaticString are as follows.

StaticString{N} is just a wrapper of a NTuple{N,UInt8} of exactly N codeunits, padded with \0, nul, bytes.

ShortStaticString{N} is just a wrapper of a NTuple{N,UInt8} of up to N codeunits, where N <= 256

LongStaticString{N} is just a wrapper of a NTuple{N,UInt8} of up to N codeunits, where N <= typemax(Int)

CStaticString{N} is similar to a StaticString but requires all the NUL bytes to be terminal codeunits. The struct also contains an extra terminal NUL.

PaddedStaticString{N,PAD} is siimlar to StaticString but is padded with an arbitrary byte codeunit.

Usage

julia> using StaticStrings

julia> Static"Hello world!"
Static"Hello world!"12

julia> Static"Hello world!" |> typeof
StaticString{12}

julia> CStatic"Hello world!"
CStatic"Hello world!"12

julia> ccall(:printf, Cint, (Ptr{Cchar},), cs)
Hello world!
13

julia> ccall(:printf, Cint, (Ptr{CStaticString{13}},), Ref(cs))
Hello world!
13

julia> Padded"Hello "20
Padded"Hello "20

julia> ps = Padded"Hello "20
Padded"Hello "20

julia> StaticString(ps)
Static"Hello               "20

julia> push!(StaticString["Hello"], "Bye")
2-element Vector{StaticString}:
 Static"Hello"5
 Static"Bye"5

StaticStrings.StaticStrings

StaticStrings.AbstractStaticString

StaticStrings.CStaticString

StaticStrings.LongStaticString

StaticStrings.PaddedStaticString

StaticStrings.ShortStaticString

StaticStrings.StaticString

StaticStrings.@CStatic_str

StaticStrings.@Long_str

StaticStrings.@Padded_str

StaticStrings.@Short_str

StaticStrings.@Static_str

StaticStrings.StaticStrings — Module

StaticStrings

Implements an AbstractString based on NTuple{N,UInt8}.

See StaticString, CStaticString

StaticStrings.AbstractStaticString — Type

AbstractStaticString{N}

Represents a string of N codeunits

StaticStrings.CStaticString — Type

CStaticString(data::NTuple{N,UInt8})

AbstractStaticString that stores codeunits in a NTuple{N,UInt8} but requires NUL codeunits to be at the end.

N.B. The size of a CStaticString{N} is N+1 bytes.

StaticStrings.LongStaticString — Type

LongStaticString(data::NTuple{N,UInt8}, ncodeunits::Int)

AbstractStaticString that stores ncodeunits codeunits in a NTuple{N,UInt8}

N.B. The size of a LongStaticString{N} is N+8 bytes.

StaticStrings.PaddedStaticString — Type

PaddedStaticString{N,PAD}(data::NTuple{N,UInt8})

AbstractStaticString that is padded with PAD.

StaticStrings.ShortStaticString — Type

ShortStaticString(data::NTuple{N,UInt8}, ncodeunits::UInt8)

AbstractStaticString that stores ncodeunits codeunits in a NTuple{N,UInt8}

N.B. The size of a ShortStaticString{N} is N+1 bytes.

StaticStrings.StaticString — Type

StaticString(data::NTuple{N,UInt8})

AbstractStaticString that stores codeunits in a NTuple{N,UInt8}.

StaticStrings.@CStatic_str — Macro

CStatic"string"[N]

Create a CstaticString from "string". Optionally, specify the number of codeunits, N.

StaticStrings.@Long_str — Macro

Long"string"N

Create a LongStaticString from "string". Optionally, specify the number of codeunits, N.

StaticStrings.@Padded_str — Macro

Padded"string[PAD]"N

Creat PaddedStaticString of N codeunits from "string". The last codeunit in the provided string becomes the pad.

StaticStrings.@Short_str — Macro

Short"string"N

Create a ShortStaticString from "string". Optionally, specify the number of codeunits, N.

StaticStrings.@Static_str — Macro

Static"string"N

Create a StaticString from "string". Optionally, specify the number of codeunits, N.